requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
torch>=2.1.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
pydantic>=2.0.0,<3.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
python-dateutil>=2.8.0
tqdm>=4.62.0
python-dotenv>=0.19.0

# Vector databases
pinecone>=3.0.0
faiss-cpu>=1.7.0
